Move BackToUp component to the end for improved visibility
Issue: The visibility of the BackToUp component was affected by the stacking order on the homepage.

Solution: Reordered the placement of the BackToUp component to make it  visible element on the page without adjusting z-index property, resolving the visibility issue. This change enhances the user experience by ensuring the BackToUp button is prominently displayed.

This is my first open-source contribution, and I'm open to feedback. Please review, and let me know if there are any additional adjustments required. Thank you!